- Static files have been configured for the project.  See /static from the project root. - New Source form uses a lot more Django features now, so it adheres a lot more to D.R.Y.  There's also an Edit Source page which re-uses this form. - JavaScript code has been moved to external .js files.  These are in static-file folders.  The JavaScript also uses namespacing (no global variables). - CSS code of the navigation menus has been changed a bit, to allow a second navigation menu to inherit the style of the user navigation menu.
-----

This is a combination of 12 commits.
The first commit's message is:
- Now, all that newsource.html uses for the form HTML is {{ form.as_p }}.  So that part's very DRY, though the help_text labels could be better placed... (and FieldSets would be nice).
 Next: must find a better place to put the JavaScript to make it DRY across multiple pages using the form.

This is the 2nd commit message:

- Static file includes have been configured to work.
- ImageSourceForm's JavaScript is now moved to a static .js file, in images/static/js. (/templates/static would be more for project-wide JS, I think.)

This is the 3rd commit message:

- No JS in the body of the New Source template now.  The on-load code is all in ImageSourceFormHelper.js.  This means the template now has no implementation details of the form whatsoever.
Next order of business: move general utility functions to a separate .js file available to the whole project, under a namespace: like lib.stripSpaces(), or util.stripSpaces().

This is the 4th commit message:

- Added an "Edit Source" form to the Source main page.  It reuses the "New Source" form.
- Added success and error messages to the New Source and Edit Source forms.
Next: fix a quirk in the Edit Source form where if you make a mistake in the form, then the static info section reflects the changes you made anyway (though the changes don't go in the database in this case, since you can then reenter the URL for that page and the info will revert back to what it was before.)

This is the 5th commit message:

- "Edit Source" form is now in a separate Edit Source page; I think the only good method of merging the Source main page and the Source edit page is to use AJAX.  Maybe for later.
- Now using named URLs in the images app.
Next: (1) put general-utility JavaScript functions in a file in templates/static/. (2) add more navigation: (a) home page to New Source link, (b) Cancel Changes button from Source edit to Source main page with a "Changes Cancelled" message, and (c) a navigation-bar link taking you to the Source main page from the Source home or edit pages. (More to come once this is integrated with the user system...)

This is the 6th commit message:

- Added "Cancel Changes" button to the Edit Source form.
- Added base template for the images app.  This includes a menu for a particular Source: button links for Source main page, Edit Source, etc.
- Added link to the New Source page from the website index page.
- Renamed "newsource" page to "source_new".

This is the 7th commit message:

- Static file configuration is done properly now.
- Added static folder to project root, and put js/util.js inside of it; this file has general-purpose utility functions.

This is the 8th commit message:

- Renaming source_base.html to base_source.html, apparently this follows Django convention a bit more closely. (https://docs.djangoproject.com/en/1.3/topics/templates/#template-inheritance)

This is the 9th commit message:

- We have some passable CSS going on here.
Next, I want to have some CSS classes so I can reuse code between the User-related buttons and the Source-related buttons.

This is the 10th commit message:

- CSS classes used for the navigation menus now.  So the User menu and Source menu CSS code differs in only a few lines of color attributes.

This is the 11th commit message:

- The JavaScript code now uses some form of namespacing.

This is the 12th commit message:

Committing .pyc's.