Sources model now has Permissions (managed automatically by Guardian) and a Description field.
For now, there's just one permission ('all') and permissions are for a particular user and source.

How the site logic has changed: login is required to create a Source.  Users can only Edit sources that they have created.

Also, minor tweaks to Source templates and CSS.

-----

This is a combination of 3 commits.
The first commit's message is:
- Added an optional description field to Sources.  Also figured out how to properly display a text field with newlines (a Django template filter called linebreaks).

This is the 2nd commit message:

- Creating Sources now requires logging in.
- Added per-user permissions for Sources.  Currently, there's just one kind of permission: 'all'.  The effect for now is that users can only Edit the Sources that they have created; if they try to Edit another Source, they'll be prompted to log in.  (This "per-object permissions" concept isn't supported by Django out of the box; Django only has model-wide permissions like "Can edit any Source".  But with Guardian, per-object permissions are relatively simple.)
- Tweaked templates and CSS a bit for Source pages.

This is the 3rd commit message:

Committing .pyc's.